+++ src/HorseCollisionMod/Marks.lua
@@ -29,7 +29,7 @@
 --
 -- @module HorseCollisionMod.Marks
 -- @author jrandall54
--- @release 4.9.0
+-- @release 4.9.1
 
 --- Body zones bloodied for each impact direction.
 --